requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
pyodbc==5.2.0
python-tds==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlalchemy_pytds==1.0.2
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0

# RDF and Graph Database
rdflib==7.0.0
SuRF==1.3.0
sparqlwrapper==2.0.0

# Meilisearch
meilisearch==0.28.0

# AWS
boto3==1.34.0
botocore==1.34.0

# Additional utilities
python-multipart==0.0.9
python-dotenv==1.0.1
jwt==1.3.1
passlib==1.7.4
bcrypt==4.1.2
httpx==0.27.0
